+++ tests/test_eof.py
@@ -225,6 +225,13 @@ def test_load_s1_orb(filename):
 
 
 @pytest.mark.parametrize("filename", S1X_XXXORB)
+def test_load_s1_orb_strict(filename):
+    pytest.importorskip("lxml")
+    with pytest.raises(xseof.ParserError):
+        xseof.load(filename, strict=True)
+
+
+@pytest.mark.parametrize("filename", S1X_XXXORB)
 def test_load_s1_orb_from_string(filename):
     pytest.importorskip("lxml")
     xml_string = filename.read_text()
@@ -234,9 +241,25 @@ def test_load_s1_orb_from_string(filename):
 
 
 @pytest.mark.parametrize("filename", S1X_XXXORB)
+def test_load_s1_orb_from_string_strict(filename):
+    pytest.importorskip("lxml")
+    xml_string = filename.read_text()
+    with pytest.raises(xseof.ParserError):
+        xseof.from_string(xml_string, strict=True)
+
+
+@pytest.mark.parametrize("filename", S1X_XXXORB)
 def test_load_s1_orb_from_bytes(filename):
     pytest.importorskip("lxml")
     xml_bytes = filename.read_bytes()
     obj = xseof.from_string(xml_bytes)
     header = _get_header(obj)
     assert header.fixed_header.file_name == filename.stem
+
+
+@pytest.mark.parametrize("filename", S1X_XXXORB)
+def test_load_s1_orb_from_bytes_strict(filename):
+    pytest.importorskip("lxml")
+    xml_bytes = filename.read_bytes()
+    with pytest.raises(xseof.ParserError):
+        xseof.from_string(xml_bytes, strict=True)