+++ NEWS.rst
@@ -1,10 +1,12 @@
 Version history
 ===============
 
-xseof v1.0.1 (UNRELEASED)
+xseof v1.1.0 (UNRELEASED)
 -------------------------
 
 * Fix loading of Sentine-1 orbit form string.
+* New `strict` option (default: `False`) to enforce strict XML namespaces
+  checking in `xseof.load` and `xseof.from_string`.
 
 
 xseof v1.0.0 (20/11/2022)